integration/requirements.txt
# Core data analysis libraries
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0

# Machine learning libraries
scikit-learn>=1.1.0
xgboost>=1.6.0
lightgbm>=3.3.0

# Financial data libraries
yfinance>=0.1.87
ta-lib>=0.4.25
ta>=0.10.2

# Visualization libraries
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Web scraping and API libraries
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.5.0

# Database libraries
sqlalchemy>=1.4.0
pymysql>=1.0.0

# Utility libraries
python-dotenv>=0.20.0
click>=8.1.0
rich>=12.5.0
tqdm>=4.64.0

# Date and time handling
python-dateutil>=2.8.0
pytz>=2022.2

# Text processing and sentiment analysis
nltk>=3.7
textblob>=0.17.0
vaderSentiment>=3.3.0

# HTTP client
httpx>=0.23.0

# Configuration management
pyyaml>=6.0

# Logging
loguru>=0.6.0